Restructure node events
